+++ src/main/java/school/faang/user_service/dto/RequestStatusDto.java
@@ -1,0 +1,8 @@
+package school.faang.user_service.dto;
+
+import lombok.Data;
+
+@Data
+public class RequestStatusDto {
+    private String status;
+}
