+++ src/main/java/school/faang/user_service/filter/mentorship/MentorshipRequestDescriptionFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.mentorship;
 
 import lombok.NonNull;
 import org.springframework.stereotype.Component;
@@ -8,17 +8,17 @@ import school.faang.user_service.entity.MentorshipRequest;
 import java.util.stream.Stream;
 
 @Component
-public class MentorshipRequestStatusFilter implements MentorshipRequestFilter {
+public class MentorshipRequestDescriptionFilter implements MentorshipRequestFilter {
     @Override
     public boolean isApplicable(@NonNull RequestFilterDto requestFilterDto) {
-        return requestFilterDto.getStatus() != null;
+        return requestFilterDto.getDescription() != null
+                && !requestFilterDto.getDescription().isBlank();
     }
 
     @Override
     public Stream<MentorshipRequest> apply(
             Stream<MentorshipRequest> mentorshipRequests, RequestFilterDto requestFilterDto) {
-        return mentorshipRequests
-                .filter(mentorshipRequest
-                        -> requestFilterDto.getStatus().equals(mentorshipRequest.getStatus()));
+        return mentorshipRequests.filter(mentorshipRequest
+                -> requestFilterDto.getDescription().equalsIgnoreCase(mentorshipRequest.getDescription()));
     }
 }
