+++ src/main/java/school/faang/user_service/filter/mentorship/MentorshipRequestRequesterIdFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.mentorship;
 
 import lombok.NonNull;
 import org.springframework.stereotype.Component;
@@ -8,17 +8,18 @@ import school.faang.user_service.entity.MentorshipRequest;
 import java.util.stream.Stream;
 
 @Component
-public class MentorshipRequestDescriptionFilter implements MentorshipRequestFilter {
+public class MentorshipRequestRequesterIdFilter implements MentorshipRequestFilter {
     @Override
     public boolean isApplicable(@NonNull RequestFilterDto requestFilterDto) {
-        return requestFilterDto.getDescription() != null
-                && !requestFilterDto.getDescription().isBlank();
+        return requestFilterDto.getRequesterId() != null
+                && requestFilterDto.getRequesterId() > 0;
     }
 
     @Override
     public Stream<MentorshipRequest> apply(
             Stream<MentorshipRequest> mentorshipRequests, RequestFilterDto requestFilterDto) {
-        return mentorshipRequests.filter(mentorshipRequest
-                -> requestFilterDto.getDescription().equalsIgnoreCase(mentorshipRequest.getDescription()));
+        return mentorshipRequests
+                .filter(mentorshipRequest -> requestFilterDto.getRequesterId()
+                        .equals(mentorshipRequest.getRequester().getId()));
     }
 }
