+++ src/test/java/school/faang/user_service/filter/MentorshipRequestDescriptionFilterTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
+import school.faang.user_service.filter.mentorship.MentorshipRequestDescriptionFilter;
 
 import java.util.stream.Stream;
