+++ src/test/java/school/faang/user_service/filter/MentorshipRequestReceiverIdFilterTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.User;
+import school.faang.user_service.filter.mentorship.MentorshipRequestReceiverIdFilter;
 
 import java.util.stream.Stream;
