+++ src/test/java/school/faang/user_service/service/mentorship_request_filter_test/MentorshipRequestDescriptionFilterTest.java
@@ -2,7 +2,7 @@ package school.faang.user_service.service.mentorship_request_filter_test;
 
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
-import school.faang.user_service.filter.MentorshipRequestFilter;
+import school.faang.user_service.filter.mentorship.MentorshipRequestFilter;
 
 import java.util.stream.Stream;
