+++ src/test/java/school/faang/user_service/service/mentorship_request_filter_test/MentorshipRequestStatusFilterTest.java
@@ -3,7 +3,7 @@ package school.faang.user_service.service.mentorship_request_filter_test;
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.RequestStatus;
-import school.faang.user_service.filter.MentorshipRequestFilter;
+import school.faang.user_service.filter.mentorship.MentorshipRequestFilter;
 
 import java.util.stream.Stream;
